@@ -1,9 +1,12 @@
 package model;
 
+import lombok.Data;
+
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
-import java.util.Date;
+import java.time.LocalDateTime;
 
+@Data
 @Entity
 @Table(name = "captcha_codes")
 public class CaptchaCode {
@@ -12,7 +15,7 @@ public class CaptchaCode {
     private int id;
 
     @NotNull
-    private Date time;
+    private LocalDateTime time;
 
     @NotNull
     private String code;
@@ -20,36 +23,4 @@ public class CaptchaCode {
     @NotNull
     @Column(name = "secret_code")
     private String secretCode;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public Date getTime() {
-        return time;
-    }
-
-    public void setTime(Date time) {
-        this.time = time;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public void setCode(String code) {
-        this.code = code;
-    }
-
-    public String getSecretCode() {
-        return secretCode;
-    }
-
-    public void setSecretCode(String secretCode) {
-        this.secretCode = secretCode;
-    }
 }
