@@ -1,12 +1,15 @@
 package model;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import lombok.Data;
 import enums.ModerationStatus;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
-import java.util.Date;
+import java.time.LocalDateTime;
 import java.util.List;
 
+@Data
 @Entity
 @Table(name = "posts")
 public class Post {
@@ -15,11 +18,14 @@ public class Post {
     private int id;
 
     @NotNull
-    private byte is_active;
+    @Column(name = "is_active")
+    @JsonIgnore
+    private byte isActive;
 
     @Enumerated(EnumType.STRING)
     @NotNull
     @Column(columnDefinition = "moderation_status DEFAULT 'NEW'")
+    @JsonIgnore
     private ModerationStatus moderationStatus;
 
 
@@ -28,7 +34,7 @@ public class Post {
     private User user;
 
     @NotNull
-    private Date time;
+    private LocalDateTime time;
 
     @NotNull
     private String title;
@@ -47,75 +53,6 @@ public class Post {
     )
     private List<Tag> postTags;
 
-    public List<Tag> getPostTags() {
-        return postTags;
-    }
-
-    public void setPostTags(List<Tag> postTags) {
-        this.postTags = postTags;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public byte getIs_active() {
-        return is_active;
-    }
-
-    public void setIs_active(byte is_active) {
-        this.is_active = is_active;
-    }
-
-    public ModerationStatus getModerationStatus() {
-        return moderationStatus;
-    }
-
-    public void setModerationStatus(ModerationStatus moderationStatus) {
-        this.moderationStatus = moderationStatus;
-    }
-
-    public User getUser() {
-        return user;
-    }
-
-    public void setUser(User user) {
-        this.user = user;
-    }
-
-    public Date getTime() {
-        return time;
-    }
-
-    public void setTime(Date time) {
-        this.time = time;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getText() {
-        return text;
-    }
-
-    public void setText(String text) {
-        this.text = text;
-    }
-
-    public int getViewCount() {
-        return viewCount;
-    }
-
-    public void setViewCount(int viewCount) {
-        this.viewCount = viewCount;
-    }
+    @OneToMany(mappedBy = "post")
+    private List<PostVote> postVote;
 }
